quality of life improvements

+++ src/initialize.jl
@@ -19,7 +19,7 @@ function build(instance::String; dir=joinpath(dirname(@__DIR__), "instances"))
     for k ∈ K N[k] = Node(df[k,1], df[k,2], df[k,3]) end
     # Arcs
     df = DataFrame(CSV.File(joinpath(dir, "$instance/arcs.csv"), header=false))
-    A  = Dict{Tuple{Int64,Int64},Arc}()
+    A  = Dict{Tuple{Int,Int},Arc}()
     for i ∈ K for j ∈ K A[(i,j)] = Arc(i, j, df[i,j]) end end
     G  = (N, A)
     return G
@@ -41,27 +41,28 @@ Note, `dir` locates the the folder containing instance files as sub-folders.
         |-arcs.csv
 """
 function savings(instance::String; dir=joinpath(dirname(@__DIR__), "instances"))
+    # Step 1: Initialize
     G = build(instance; dir=dir)
     s = Solution(G...)
     N = s.N
     d = N[1]
-    # Step 1: Initialize solution with routes to every node from the depot node (first node)
+    # Step 2: Initialize solution with routes to every node from the depot node (first node)
     k = length(N)
     K = eachindex(N)
     D = [deepcopy(d) for _ ∈ K]
     for k ∈ reverse(K) insertnode!(N[k], D[k], D[k], s) end
-    # Step 2: Merge routes iteratively until single route traversing all nodes remains
+    # Step 3: Merge routes iteratively until single route traversing all nodes remains
     X = fill(-Inf, (K,K))       # X[i,j]: Savings from merging route with tail node N[i] into route with tail node N[j]
-    ϕ = ones(Int64, K)          # ϕ[k]  : binary weight for route k
+    ϕ = ones(Int, K)            # ϕ[k]  : binary weight for route k
     for _ ∈ 3:k
-        # Step 2.1: Iterate through every route-pair combination
+        # Step 3.1: Iterate through every route-pair combination
         z = f(s)
         for i ∈ 2:k
             n₁ = N[i]
             d₁ = D[i]
             if !isequal(n₁.h, d₁.i) continue end
             for j ∈ 2:k
-                # Step 2.1.1: Merge routes with tail node N[i] into route with tail node N[j]
+                # Step 3.1.1: Merge routes with tail node N[i] into route with tail node N[j]
                 n₂ = N[j]
                 d₂ = D[j]
                 if isequal(n₁, n₂) continue end
@@ -78,11 +79,11 @@ function savings(instance::String; dir=joinpath(dirname(@__DIR__), "instances"))
                     nₒ = nₜ
                     nₜ = isone(nₒ.t) ? d₁ : N[nₒ.t]
                 end
-                # Step 2.1.2: Compute savings from merging route with tail node N[i] into route with tail node N[j]
+                # Step 3.1.2: Compute savings from merging route with tail node N[i] into route with tail node N[j]
                 z⁻ = f(s)
                 Δ  = z - z⁻
                 X[i,j] = Δ
-                # Step 2.1.3: Unmerge routes with tail node N[i] into route with tail node N[j]
+                # Step 3.1.3: Unmerge routes with tail node N[i] into route with tail node N[j]
                 nₜ = d₁
                 nₒ = N[n₂.h]
                 nₕ = isone(nₒ.h) ? d₂ : N[nₒ.h]
@@ -96,7 +97,7 @@ function savings(instance::String; dir=joinpath(dirname(@__DIR__), "instances"))
                 end
             end
         end
-        # Step 2.2: Merge routes that render highest savings
+        # Step 3.2: Merge routes that render highest savings
         i,j = Tuple(argmax(X))
         n₁ = N[i]
         d₁ = D[i]
@@ -114,7 +115,7 @@ function savings(instance::String; dir=joinpath(dirname(@__DIR__), "instances"))
             nₜ = isone(nₒ.t) ? d₁ : N[nₒ.t]
         end
         D[i] = D[j]
-        # Step 2.3: Revise savings and selection vectors appropriately
+        # Step 3.3: Revise savings and selection vectors appropriately
         X[i, :] .= -Inf
         X[:, i] .= -Inf
         X[j, :] .= -Inf
@@ -123,7 +124,7 @@ function savings(instance::String; dir=joinpath(dirname(@__DIR__), "instances"))
     end
     for n ∈ N d.t = isequal(n.h, d.i) ? n.i : d.t end
     for n ∈ N d.h = isequal(n.t, d.i) ? n.i : d.h end
-    # Step 3: Return initial solution
+    # Step 4: Return initial solution
     return s
 end
